Updated parsing of excluded caller id list
The parsing of each value in the list of caller ids to exclude was being handled by ord(char), which does not accept
strings with multiple characters. Since some caller ids do have multiple digits, it makes more sense to attempt to cast
the string to an int to parse the ids (if an int is not included in the string, the cast throws an exception, which is caught).