Remove output of callers with no VCF in a sample
This commit removes the output (from both console and all aspects of the output directory) of
all callers that are not used in a sample. Near the beginning of execution, 'scallers' is created,
which is a dictionary that has each sample as a key and the caller dictionary for that sample as a value.
Warnings are given when there are no usable VCFs in the sample directory (due to wrong sample IDs, or every sample ID is excluded with -E argument).
When the model is being applied to samples, only the callers that had VCFs in that sample will be used to generate
the output data for cross_fold_stats and detailed_stats.

The read_partitions_by_caller method in fusor_utils.py has been adjusted so that the returned data structure
only has entries at the caller level for callers that had partitions to be read. E.g. if a caller was not used in any sample,
there will be no partitions with that caller id, so there will be no entry for that caller in the returned structure. With regards
to the output directory, this removes all entries in the sim_matrix that include unused callers and arbitrary 0's for data fields.

The export_caller_by_type_and_bin method in fusor_utils.py has been adjusted so that a caller is not exported if it is not used. Previously,
all unused callers were looped through and exported with arbitrary 0's as data.