Made the reference fasta a required input
The reference fasta could previously be omitted as an argument, which would later cause the program to crash.
It is now a required argument, and it is verified that the file exists at the beginning of execution.
If it is not specified or it does not exist, an IOError is raised to inform the user of the respective error.