Removed unused caller output (#16)
* Updated parsing of excluded caller id list

The parsing of each value in the list of caller ids to exclude was being handled by ord(char), which does not accept strings with multiple characters. Since some caller ids do have multiple digits, it makes more sense to attempt to cast the string to an int to parse the ids (if an int is not included in the string, the cast throws an exception, which is caught).

* Improved handling of stage_exclude_list argument

Added a check to the stage_exclude_list argument to make sure every id the
user wants to exclude is a valid caller id. Also improved clarity of error
messages regarding this argument.

* Removed excluded callers as a parameter from several functions

There were a handful of functions that took the callers dictionary and the
list of excluded callers as parameters, and then found the set difference between them
(all non-excluded callers) to loop through to perform the function's goal. Since excluded
callers were removed from the callers dictionary in a pervious commit, finding this set difference
is no longer necessary, nor is the parameter for excluded callers.

* Remove output of callers with no VCF in a sample

This commit removes the output (from both console and all aspects of the output directory) of
all callers that are not used in a sample. Near the beginning of execution, 'scallers' is created,
which is a dictionary that has each sample as a key and the caller dictionary for that sample as a value.
Warnings are given when there are no usable VCFs in the sample directory (due to wrong sample IDs, or every sample ID is excluded with -E argument).
When the model is being applied to samples, only the callers that had VCFs in that sample will be used to generate the output data for cross_fold_stats and detailed_stats.

The read_partitions_by_caller method in fusor_utils.py has been adjusted so that the returned data structure only has entries at the caller level for callers that had partitions to be read. E.g. if a caller was not used in any sample, there will be no partitions with that caller id, so there will be no entry for that caller in the returned structure. With regards to the output directory, this removes all entries in the sim_matrix that include unused callers and arbitrary 0's for data fields.

The export_caller_by_type_and_bin method in fusor_utils.py has been adjusted so that a caller is not exported if it is not used. Previously, all unused callers were looped through and exported with arbitrary 0's as data.